tripo-mcp/pyproject.toml
[project]
name = "tripo-mcp"
version = "0.1.0"
description = "Tripo3D API 本地工具集成，支持文本/图片/多视图转3D、贴图、动画、风格化、格式转换、余额查询等。"
authors = [
    { name="lxy2109", email="cinnamon_l@163.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "pydantic",
    "python-dotenv",
    "httpx",
    "mcp[cli]"
]

[project.scripts]
tripo-mcp = "src.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"